+++ dart/geobase/CHANGELOG.md
@@ -6,6 +6,7 @@ NOTE: Version 0.6.0 currently [under development](https://github.com/navibyte/ge
 * [Add copyWith (and map) methods to coordinate, meta, feature and geometry classes #189](https://github.com/navibyte/geospatial/issues/189)
 * [Populating and unpopulating bounds in geometries and feature objects #197](https://github.com/navibyte/geospatial/issues/197)
 * [Testing feature and geometry equality #188](https://github.com/navibyte/geospatial/issues/188)
+* [Factories on Position and Box to create from double arrays #199](https://github.com/navibyte/geospatial/issues/199)
 
 ⚠️ Breaking changes:
 * [Allow a position stored as "Position" in Point, and a bounding box as "Box" in Geometry classes #192](https://github.com/navibyte/geospatial/issues/192)